# Optimize RAM usage and performance by removing redundant `rawData`