--- a/ZAPD/StringHelper.h
+++ b/ZAPD/StringHelper.h
@@ -7,6 +7,14 @@
 #include <vector>
 #include <algorithm>
 
+#ifndef __PRETTY_FUNCTION__
+#ifdef _MSC_VER
+#define __PRETTY_FUNCTION__ __FUNCSIG__
+#else
+#define __PRETTY_FUNCTION__ __func__
+#endif
+#endif
+
 class StringHelper
 {
 public: