--- a/ZAPD/StringHelper.h
+++ b/ZAPD/StringHelper.h
@@ -1,11 +1,11 @@
 #pragma once
 
+#include <algorithm>
 #include <cstring>
 #include <numeric>
 #include <stdarg.h>
 #include <string>
 #include <vector>
-#include <algorithm>
 
 #ifndef __PRETTY_FUNCTION__
 #ifdef _MSC_VER
@@ -109,7 +109,7 @@ public:
 
 	static std::string BoolStr(bool b) { return b ? "true" : "false"; }
 
-	static bool HasOnlyDigits(const std::string &str)
+	static bool HasOnlyDigits(const std::string& str)
 	{
 		return std::all_of(str.begin(), str.end(), ::isdigit);
 	}