--- a/ZAPD/ZBackground.cpp
+++ b/ZAPD/ZBackground.cpp
@@ -19,7 +19,7 @@ ZBackground::ZBackground(const std::string& prefix, const std::vector<uint8_t>& 
                          uint32_t nRawDataIndex, ZFile* nParent)
 	: ZResource(nParent)
 {
-	rawData.assign(nRawData.begin(), nRawData.end());
+	
 	rawDataIndex = nRawDataIndex;
 	name = GetDefaultName(prefix.c_str(), rawDataIndex);
 	outName = name;
@@ -31,6 +31,7 @@ void ZBackground::ParseRawData()
 {
 	ZResource::ParseRawData();
 
+	const auto& rawData = parent->GetRawData();
 	size_t i = 0;
 	while (true)
 	{
@@ -61,10 +62,10 @@ void ZBackground::ParseBinaryFile(const std::string& inFolder, bool appendOutNam
 	CheckValidJpeg(filepath.generic_string());
 }
 
-void ZBackground::ExtractFromXML(tinyxml2::XMLElement* reader, const std::vector<uint8_t>& nRawData,
+void ZBackground::ExtractFromXML(tinyxml2::XMLElement* reader,
                                  uint32_t nRawDataIndex)
 {
-	ZResource::ExtractFromXML(reader, nRawData, nRawDataIndex);
+	ZResource::ExtractFromXML(reader, nRawDataIndex);
 	DeclareVar("", "");
 }