--- a/ZAPD/ZBackground.h
+++ b/ZAPD/ZBackground.h
@@ -11,8 +11,9 @@ protected:
 
 public:
 	ZBackground(ZFile* nParent);
-	ZBackground(const std::string& prefix, const std::vector<uint8_t>& nRawData,
+	ZBackground(const std::string& prefix,
 	            uint32_t nRawDataIndex, ZFile* nParent);
+
 	void ParseRawData() override;
 	void ParseBinaryFile(const std::string& inFolder, bool appendOutName);
 	void ExtractFromXML(tinyxml2::XMLElement* reader,