--- a/ZAPD/ZCutscene.h
+++ b/ZAPD/ZCutscene.h
@@ -434,8 +434,8 @@ public:
 
 	ZResourceType GetResourceType() const override;
 
-	void ExtractFromXML(tinyxml2::XMLElement* reader, const std::vector<uint8_t>& nRawData,
-	                    const uint32_t nRawDataIndex) override;
+	void ExtractFromXML(tinyxml2::XMLElement* reader,
+	                    uint32_t nRawDataIndex) override;
 
 protected:
 	int32_t numCommands;