--- a/ZAPD/ZCutscene.h
+++ b/ZAPD/ZCutscene.h
@@ -434,8 +434,7 @@ public:
 
 	ZResourceType GetResourceType() const override;
 
-	void ExtractFromXML(tinyxml2::XMLElement* reader,
-	                    uint32_t nRawDataIndex) override;
+	void ExtractFromXML(tinyxml2::XMLElement* reader, uint32_t nRawDataIndex) override;
 
 protected:
 	int32_t numCommands;