--- a/ZAPD/ZDisplayList.h
+++ b/ZAPD/ZDisplayList.h
@@ -346,11 +346,10 @@ public:
 	std::vector<uint32_t> references;
 
 	std::string defines;  // Hack for special cases where vertex arrays intersect...
-	std::vector<uint8_t> fileData;
 	std::vector<ZMtx> mtxList;
 
 	ZDisplayList(ZFile* nParent);
-	ZDisplayList(std::vector<uint8_t> nRawData, uint32_t rawDataIndex, int32_t rawDataSize,
+	ZDisplayList(uint32_t rawDataIndex, int32_t rawDataSize,
 	             ZFile* nParent);
 	~ZDisplayList();
 
@@ -362,7 +361,7 @@ public:
 	Declaration* DeclareVar(const std::string& prefix, const std::string& bodyStr);
 
 	void TextureGenCheck(std::string prefix);
-	static bool TextureGenCheck(std::vector<uint8_t> fileData, ZRoom* scene, ZFile* parent,
+	static bool TextureGenCheck(ZRoom* scene, ZFile* parent,
 	                            std::string prefix, int32_t texWidth, int32_t texHeight,
 	                            uint32_t texAddr, uint32_t texSeg, F3DZEXTexFormats texFmt,
 	                            F3DZEXTexSizes texSiz, bool texLoaded, bool texIsPalette,
@@ -384,5 +383,5 @@ public:
 	ZResourceType GetResourceType() const override;
 
 protected:
-	std::vector<uint8_t> dlistRawData;
+	size_t numInstructions;
 };