--- a/ZAPD/ZFile.cpp
+++ b/ZAPD/ZFile.cpp
@@ -194,7 +194,7 @@ void ZFile::ParseXML(ZFileMode mode, XMLElement* reader, std::string filename, b
 			ZResource* nRes = nodeMap[nodeName](this);
 
 			if (mode == ZFileMode::Extract)
-				nRes->ExtractFromXML(child, rawData, rawDataIndex);
+				nRes->ExtractFromXML(child, rawDataIndex);
 
 			auto resType = nRes->GetResourceType();
 			if (resType == ZResourceType::Texture)