… ZAPD/ZLimb.h
… ZAPD/ZMtx.cpp
… ZAPD/ZMtx.h
… ZAPD/ZPath.cpp
… ZAPD/ZPath.h
… ZAPD/ZResource.cpp
… ZAPD/ZResource.h
… ZAPD/ZRoom/Commands/SetAnimatedMaterialList.cpp
… ZAPD/ZRoom/Commands/SetCollisionHeader.cpp
… ZAPD/ZRoom/Commands/SetCsCamera.cpp
… ZAPD/ZRoom/Commands/SetCutscenes.cpp
… ZAPD/ZRoom/Commands/SetMesh.cpp
… ZAPD/ZRoom/Commands/SetMesh.h
… ZAPD/ZRoom/ZRoom.cpp
… ZAPD/ZRoom/ZRoom.h
… ZAPD/ZScalar.cpp
… ZAPD/ZSkeleton.cpp
… ZAPD/ZSkeleton.h
… ZAPD/ZString.cpp
… ZAPD/ZSymbol.cpp
… ZAPD/ZSymbol.h
… ZAPD/ZTexture.cpp
… ZAPD/ZTexture.h
… ZAPD/ZVector.cpp
… ZAPD/ZVtx.cpp