Currently each `ZResource` has the vector member `rawData`, which contains a copy of the binary data of the whole file being extracted (ie an object, scene, etc). This means that it makes a copy of the whole data vector once for every resource, using an unnecessary amount of memory and CPU time from all these redundant copies.
This solution removes said member, and instead makes every resource ask for a reference of the data to their `parent` (`ZFile`).

To put this change in perspective, here are some numbers¹:

| Measuring tool | Current master | This PR | % |
|--------------------|------------------------|--------------------|--------------------|
| Run time (`time ./extract_assets.py`)                | `0m23.732s` | `0m5.079s` | `−78.62%` |
| CPU cycles (`perf record ./extract_assets.py`)   | `179.61 GCycles` (or `179,606,898,977 cycles`) | `25.98 GCycles` (or `25,981,576,639 cycles`) | `-85.53%` |
| RAM usage² (`valgrind --tool=massif`)            |  `1421.04 MiB` (or `1,421,040,336 bytes`) | `12.82 MiB` (or `13,447,320 bytes`) | `−99.05%` |


--------------------------------------
¹ Only tested with OoT repo.
² I tested this method only with `gameplay_keep`, because I couldn't find an easy and quick way to automate this for every asset we have.